wiki_sync_service/requirements.txt
requests==2.32.3
python-dotenv==1.0.0
minio==7.2.3

